feat: extract sunlight out of sky manager to allow using shadows without it
